Test 3, 2.0
Additions:

-Added some of the witchcraft table's artifacts, this system is based on a calculation based on the quality of the items given,
the items compatible with the witchcraft table are:
Cloth, Broken Talisman, Mysterious dust and Ancient Tablet. Being of quality 1, 2, 3 and 4
respectively, and also each given by a different tribesmen.

-Added tribesmen trading lootables, the trading has changed slightly, now higher tier items such as Ancient Tablets are locked away behind a progression
system with the tribe, you must now get their mask trough trading to gain access to trades of other tribesmen, for example if you don't have a mask
you'll be unable to trade with any tribesmen other than a marauder, this system works on the following order:

(Tribesmen -> mask tier required) Marauder -> 0 , Hunter -> 1, Scorcher -> 2, Spiritweaver -> 3
(Mask (tier)) None (0), Marauder (1), Hunter(2), Scorcher(3), Spiritweaver(4)

This system and the lootables are gonna be expanded upon in a future alpha.

-Added many artifacts' functionality*, some effects are triggered by tick, when damaged, when attacking or in many other ways, most of them work on the
offhand but there are a few exceptions. There are no tooltips as of now, but worry not they'll be added in the future.

*as of now only feather and irongrass are non funcional

Fixes:

-Fixed Witchcraft Table Crash
-Fixed cactus broth stack issue (actually being stackable x64 was an overlook on my end).
-Fixed some bugs related to the frost wand, now it should be more accurate.
-Fixed an issue related to the bestiarium showing a wrong image on the ghost page.
-Fixed artifacts having a max stack size of 0 (could randomly crash the game)

Tweaks:

-Retextured many old items such as sandmonster scale and desert robes.
-Added sounds for Rifted enderman's shockwave attack.
-Vessel (item) has now nbt data which means that the vessel you get (either steve or alex) will always be the same when using their item form.
-Bestiarium entity information logic has been polished.